% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMulRegressionExp.R
\name{runMulRegressionExp}
\alias{runMulRegressionExp}
\title{Function to run multiple GLMs}
\usage{
runMulRegressionExp(signatures, exposome, covariates, family)
}
\description{
Function to run multiple GLMs of the form signature ~ exposure.
}
