% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMulRegressionSig.R
\name{runMulRegressionSig}
\alias{runMulRegressionSig}
\title{Function to run multiple GLMs}
\usage{
runMulRegressionSig(signatures, exposome, formula, family)
}
\description{
Function to run multiple GLMs of the form signature ~ condition.
}
