% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRegressionSig.R
\name{runRegressionSig}
\alias{runRegressionSig}
\title{Function to run single GLM}
\usage{
runRegressionSig(dat, formula, family)
}
\description{
Function to run single GLM of the form condition ~ signature.
}
